+++ fe/deckbop/src/main.js
@@ -1,27 +1,12 @@
 import Vue from 'vue'
-import VueRouter from 'vue-router'
 import App from './App.vue'
+import axios from 'axios'
 
-// Import components here for routes
-import Register from './components/authentication/Register.vue'
-import Login from './components/authentication/Login.vue'
-import ViewDecks from './components/deckManagement/ViewDecks.vue'
-
-import store from './store/store';
+import router from './router/router'
+import store from './store/store'
 
 Vue.config.productionTip = false
-Vue.use(VueRouter)
-
-const routes = [
-  {path: '/login', component: Login},
-  {path: '/register', component: Register},
-  {path: '/viewDecks', component: ViewDecks},
-  {path: '/', component: ViewDecks},
-]
-
-const router = new VueRouter({
-  routes
-})
+Vue.prototype.$http = axios;
 
 new Vue({
   router,
